+++ src/components/functional/toasts-list/components/transaction-toast/components/transaction-toast-content/transaction-toast-content.tsx
@@ -90,11 +90,11 @@ export class TransactionToastContent {
           </div>
 
           {hasCloseButton && (
-            <fa-icon
+            <mvx-fa-icon
               icon={faTimes}
               class="transaction-toast-close-icon"
               onClick={this.handleDeleteToast.bind(this)}
-            ></fa-icon>
+            />
           )}
 
           {!hasCloseButton && showExplorerLinkButton && (