+++ src/components/functional/toasts-list/toast-list.types.ts
@@ -5,4 +5,6 @@ export enum ToastEventsEnum {
   TRANSACTION_TOAST_DATA_UPDATE = 'TRANSACTION_TOAST_DATA_UPDATE',
   // Event to update the custom toast data
   CUSTOM_TOAST_DATA_UPDATE = 'CUSTOM_TOAST_DATA_UPDATE',
+  HIDE = 'HIDE_TOAST_LIST',
+  SHOW = 'SHOW_TOAST_LIST',
 }