+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - [Updates on unlock panel and install links.](https://github.com/multiversx/mx-sdk-dapp-ui/pull/153)
+- [Added Check your device text on ledger connect](https://github.com/multiversx/mx-sdk-dapp-ui/pull/158)
+- [Fixed close button on transaction toast.](https://github.com/multiversx/mx-sdk-dapp-ui/pull/157)
+- [Fixed issue with xPortal back button on login screen](https://github.com/multiversx/mx-sdk-dapp-ui/pull/156)
+- [Fixed transaction toast loading](https://github.com/multiversx/mx-sdk-dapp-ui/pull/155)
 
 ## [[0.0.11](https://github.com/multiversx/mx-sdk-dapp-ui/pull/152)] - 2025-07-08